Moves utility code to a separate utility folder
