+++ flow/include/dispatch.h
@@ -13,7 +13,7 @@
 #include <utility>
 
 // Flow
-#include <flow/impl/static_assert.hpp>
+#include <flow/utility/static_assert.hpp>
 
 namespace flow
 {
