+++ flow/include/utility/implement_crtp_base.hpp
@@ -4,8 +4,8 @@
  *
  * @file implement_crtp_base.hpp
  */
-#ifndef FLOW_IMPL_IMPLEMENT_CRTP_BASE_H
-#define FLOW_IMPL_IMPLEMENT_CRTP_BASE_H
+#ifndef FLOW_UTILITY_IMPLEMENT_CRTP_BASE_HPP
+#define FLOW_UTILITY_IMPLEMENT_CRTP_BASE_HPP
 
 /**
  * @brief Macro used to implement CRTP-derived type helpers in a uniform way
@@ -15,4 +15,4 @@
   inline DerivedT* derived() { return static_cast<DerivedT*>(this); }                                                  \
   inline const DerivedT* derived() const { return static_cast<const DerivedT*>(this); }
 
-#endif  // FLOW_IMPL_IMPLEMENT_CRTP_BASE_H
+#endif  // FLOW_UTILITY_IMPLEMENT_CRTP_BASE_HPP
